a function that creates a file.

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@
  * Return: length of the string
  */
 
-size_t _strlen(char *str)
+int _strlen(char *str)
 {
 	size_t i;
